Update Grid: Added partial2full
Added a function to introduce limtied support for partial grids. The partial2full function is intended to convert quantities defined on a partial grid, generally referred ot as x in this program, to the equivalent quantity on the full grid by filling missing entries with zero. This is sufficient to pair with existing functions for similar functionality as full grid.