Developer (#18) (#19)
* Update README.md

* Update to overlay* functions

This includes moving overlay_line from the Grid class to the tools package, to be consistent with other overlay functions.
+ Bug fixes in some of the overlay functions.

* Updated Phantom.fit

This function can now take the vec1 and vec2 vectors (lists of elements) directly. This allows compatibility with ungridded experimental data.

* Updated n_B in get_setpoint

+ Other updateds to get_setpoint.
+ Added contourf functionality to plot2d functions.
+ Update to mass-mobility relation parameters in io.import_c.
+ Added fblue colormap.
+ Minor README updates.

* Minor update to io.import_c.m

* Update import_c.m

* Bug fixes to Grid

+ Bug fix to mass2rho
+ Added forange colormap

* Added lower cut for partial grids

* Rename prop_pma.Dm and prop_pma.rho0

This is associated with syncing the program with updates to the tfer_pma package. Default mass-mobiltiy relations used in defining the PMA setpoint are now specifies in the prop_pma functions instead of in the mp2zp function.
+ Minor updates to import_c.
+ Minor bug fix in Grid class.

* Removed Dm and rho0 defn. in import

+ Update to .gitignore

* Update gen_grid_c2.m with prop_pma.Dm name change

* Intro. varargin input to overlay*.m

These functions now take a varaible number of input which is eventually passed to the plot function to allow for line specifications.
+ Added custom legend to plot2d_scatter.m.
+ Commenting updates to gen*.m functions.

* Bug fix in overlay_isorho.m

The function was missing a pi/6 factor such that lines were plotted incorrectly.

* Updates to f* colormaps

+ Updated .gitignore

* Update plot2d_marg_b.m

Based on sugggestions by @ArashNaseri.

* Updated marginalize functions

Added marginalize_op function to generate marginalization operators.
+ Updated both marginalize functions to include partial elements (where only part of the element lies below the cut line).

* Bug fix to plot2d_marg_b

* Bug fix for Grid.ray_sum

This bug fix only applies for partial grids.
+ Added 1D kernel generation for DMA.

* Bug fixes to Grid.dr function

* Update overlay_ellipse.m

* Reverted plot2d mod based on element size

* Update to io.import_a

+ Increased the number of contours.
+ Added fgreen colormap.

* Added Phantom.fit2 method

This method allows for fitting of multimodal distributions to data/reconstructions.
+ Separated both fit methods from the single Phantom class definition. The methods are now externally defined to make them more apparent in the file tree. This is also to emphasize these methods as alternate routes to make Phantom objects (i.e. they create an instance of the Phantom class without requiring the user to call the constructor).
+ Minor updates to the rest of the Phantom class.

* Updated Phantom.mass2rho method for multimodal

* Added tools for fitting distributions

Specifically, modified Phantom.fit2 in a continued effort to improve multimodal fitting.
+ Added `logr0` arguement to Phantom.fit
+ Added sin(corr) durign fitting in an attempt to better span the space between -1 and 1

* Added tools.overlay_phantom_marg

Overlays phantom ellipses for multimodal phantoms and adds marginal distributions to appropriate panels. This function expects a figure previously produced by `Grid.plot2d_marg`.
+ Updates to inputs of tools.overlay_phantom to allow for multimodal cases.
+ Updates to commenting in tools.overlay_phantom.

* Bug fixes to Phantom.fill_p + REAMDE update

README updates consistute a sweeping overhaul of the Phantom class description, which now includes more detail and code details.

* Improved Phantom.fill_p method

Now can take rhog or mg and smd or sm.
+ Further updates to the description of the Phantom class in the README.

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Update README.md

* Limited support for uncertainty analysis of Phantom fitting

+ Added tools.redistribute

* Update to colormaps

Added CMasher colormaps.

* Updates to commenting and headers

* Moved fit2 to fit2_rho

The Phantom.fit2_rho method is tuned for effective density-mobility distributions
+ New Phantom.fit method tuned for mass-mobility distributions
+ Updated Phantom.eval* methods to take more flexible or clear arguments
+ Working updates

* Updated Phantom for mode weighting

The Phantom class can now take an additional arguement, `w`, that will rewieght the modes in the Phantom. This allows for appropriate Phantom instances when using Phantom.fit2*.

* Added Phantom.plus method

Users can now use the `+` operator to add two phantoms together.

* Updated chi > err in run_inversions*

This better reflects the Euclidean error nature of this quantity.

* Added Dm, l1, and l2 output from cov2p

* Updated `err` or `chi` -> `eps`

This variable name change is consistent with the this quantity uniformly representing the Euclidean error (Euclidean distance to exact solution).

* Update to display from Phantom.fit*

* Working updates

* Update run_inversions_g.m

* Working updates

* Update README.md

* Adjusted mu and Sigma handling in Phantom

The mu and Sigma variables are no longer cells for distributions with more than one mode. Instead, properties are appended in multidimensional arrays, e.g. Sigma is now a 3D array with the covariance in the first two dimensions, which are stacked for different modes in the third dimension.
* Note, this makes this implementation much closer to the built-in Gaussian mixture model class in MATLAB.

* Added Phantom.fit_gmm

+ Minor updates to Phantom class following previous commit.
+ Working updates.

* Added tools.mrbc2frac

Function converts mrBC-mp distributions to fraction mrBC-mp distributions.
+ Added tools.plot2d_patch
+ Merged Grid.plot2d_sweep* methods

* Colour bug fix in Grid.plot2d_sweep

Also applies the same bug fix to tools.plot2d_patch.

* Added alpha opt. to plot2d_patch

The option to added alpha is included as commented code.